new features

+++ XFCovidTrack/XFCovidTrack/Models/GlobalCases.cs
@@ -16,8 +16,8 @@ namespace XFCovidTrack.Models
         public int recovered { get; set; }
         public int active { get; set; }
         public int critical { get; set; }
-        public int casesPerOneMillion { get; set; }
-        public float deathsPerOneMillion { get; set; }
+        public double casesPerOneMillion { get; set; }
+        public double deathsPerOneMillion { get; set; }
         public int tests { get; set; }
         public float testsPerOneMillion { get; set; }
         public int affectedCountries { get; set; }
